+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Snackbar component
 * Toggle button component
 * Multiple files selection with image picker
+* Date input component
 
 ### Changed